requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pyserial==3.5
sqlalchemy==2.0.25
python-dotenv==1.0.1
pydantic==2.5.3
python-multipart==0.0.6
requests==2.31.0